pyproject.toml
[tool.poetry]
name = "llm-rpg"
version = "0.1.0"
description = ""
authors = ["WoutVossen <vossen.w@hotmail.com>"]
readme = "README.md"
packages = [{include = "capstoneProj", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
openai = "^1.57.1"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.2"
ollama = "^0.5.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pre-commit = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
